+++ lib/js_routes/middleware.rb
@@ -1,0 +1,36 @@
+class JsRoutes
+  # A Rack middleware that automatically updates routes file
+  # whenever routes.rb is modified
+  #
+  # Inspired by
+  # https://github.com/fnando/i18n-js/blob/main/lib/i18n/js/middleware.rb
+  class Middleware
+    def initialize(app)
+      @app = app
+      @routes_file = Rails.root.join("config/routes.rb")
+      @mtime = nil
+    end
+
+    def call(env)
+      update_js_routes
+      @app.call(env)
+    end
+
+    protected
+
+    def update_js_routes
+      new_mtime = routes_mtime
+      unless new_mtime == @mtime
+        JsRoutes.generate!
+        JsRoutes.definitions!
+        @mtime = new_mtime
+      end
+    end
+
+    def routes_mtime
+      File.mtime(@routes_file)
+    rescue Errno::ENOENT
+      nil
+    end
+  end
+end